// .old_project /packages/cyoda-ui-lib/src/types/types.ts
export interface ColumnInfo {
  columnName: string;
  columnPath: string;
  type: string;
  clazzType: string;
  decision: null;
  declaredClass: string;
  subClasses: string[];
  value: string | boolean | number;
}
// ...
export interface Entity {
  columnInfo: ColumnInfo;
  value: string;
  type: string;
  decision: string;
  presented?: boolean;
}